NABI-184--refactor : DTO 명 변경
